service file split and video view fix
