+++ app/views/hyrax/file_sets/media_display/_video.html.erb
@@ -1,39 +1,33 @@
 <%# app/views/hyrax/file_sets/media_display/_video.html.erb %>
-<div class="viewer-wrapper" data-turbolinks="false">
+<% content_for(:head) do %>
+    <meta name="turbolinks-visit-control" content="reload">
+<% end %>
+
+<div class="viewer-wrapper">
     <video id="video" controls class="video-js vjs-default-skin vjs-fill" style="width:100%;height:100%" preload="auto">
         <% @presenter.stored_derivatives.each do |derivative| %>
             <source src="<%= s3_url(derivative) %>" type="video/mp4" label="<%= get_derivative_res(derivative) %>" />
         <% end %>
         <%= t('hyrax.file_set.show.downloadable_content.video_tag_not_supported') %>
     </video>
+
+    <script>
+        // disable amd loader for videojs scripts
+        define._oldAmd = define.amd;
+        define.amd = null;
+    </script>
+    <script src="https://vjs.zencdn.net/8.10.0/video.js" defer></script>
+    <script src="https://unpkg.com/@silvermine/videojs-quality-selector/dist/js/silvermine-videojs-quality-selector.min.js" defer></script>
     <script>
         document.addEventListener('turbolinks:load', function (_event) {
-            if (!event.data.timing.visitStart) {
-                // load remote videojs source
-                var videoJsUrls = [
-                'https://vjs.zencdn.net/8.10.0/video.min.js',
-                'https://unpkg.com/@silvermine/videojs-quality-selector/dist/js/silvermine-videojs-quality-selector.min.js'
-                ];
-            
-                // disable amd loader for videojs scripts and append them to the head ourselves
-                define._oldAmd = define.amd;
-                define.amd = null;
-            
-                videoJsUrls.forEach(url => {
-                var script = document.createElement('script');
-                script.src = url;
-                document.head.appendChild(script);
-                })
-            
-                videojs('video', {}, function () {
+            videojs('video', {}, function () {
                 var player = this;
                 player.controlBar.addChild('QualitySelector');
-                });
-            
-                // restore the old amd loader
+            });
+
+            // restore the old amd loader
+            if (define._oldAmd) {
                 define.amd = define._oldAmd
-            } else {
-                
             }
         });
     </script>
